requirements.txt
requests==2.26.0
python-dotenv==0.18.0
progress==1.5
instagram_private_api==1.6.0.0
numpy==1.19.0
Pillow==8.3.1
extcolors==1.0.0
opencv-python==4.5.3.56
# dlib==19.22.0
pynput==1.7.2
